+++ apps/widget/modules/widget/ui/screens/widget-auth-screen/index.tsx
@@ -1,10 +1,16 @@
 import { z } from "zod"
 import { useForm } from "react-hook-form"
 import { useMutation } from "convex/react"
+import { useAtomValue, useSetAtom } from "jotai"
 import { zodResolver } from "@hookform/resolvers/zod"
 
 import { api } from "@workspace/backend/_generated/api"
 import { Doc } from "@workspace/backend/_generated/dataModel"
+
+import {
+  contactSessionIdAtomFamily,
+  organizationIdAtom,
+} from "@/modules/widget/atoms/widget-atoms"
 
 import {
   Form,
@@ -27,6 +33,11 @@ const formSchema = z.object({
 const organizationId = "123"
 
 export const WidgetAuthScreen = () => {
+  const organizationId = useAtomValue(organizationIdAtom)
+  const setContactSessionId = useSetAtom(
+    contactSessionIdAtomFamily(organizationId || "")
+  )
+
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
@@ -63,7 +74,7 @@ export const WidgetAuthScreen = () => {
       metadata,
     })
 
-    console.log({ contactSessionId })
+    setContactSessionId(contactSessionId)
   }
 
   return (