+++ apps/widget/modules/widget/ui/views/widget-view/index.tsx
@@ -3,7 +3,10 @@
 import { useAtomValue } from "jotai"
 
 import { screenAtom } from "@/modules/widget/atoms/widget-atoms"
+
 import { WidgetAuthScreen } from "@/modules/widget/ui/screens/widget-auth-screen"
+import { WidgetErrorScreen } from "@/modules/widget/ui/screens/widget-error-screen"
+import { WidgetLoadingScreen } from "@/modules/widget/ui/screens/widget-loading-screen"
 
 interface Props {
   organizationId: string
@@ -13,8 +16,8 @@ export const WidgetView = ({ organizationId }: Props) => {
   const screen = useAtomValue(screenAtom)
 
   const screenComponents = {
-    error: <p>TODO: Error</p>,
-    loading: <p>TODO: Loading</p>,
+    error: <WidgetErrorScreen />,
+    loading: <WidgetLoadingScreen organizationId={organizationId} />,
     auth: <WidgetAuthScreen />,
     voice: <p>TODO: Voice</p>,
     inbox: <p>TODO: Inbox</p>,